==> src/1.2_defining_parameters.ipynb <==
# Finding parameters

In order to constrain the model using photons, osmolarity, and CO2 intake, it is necessary to have the required parameters

import pandas as pd

modelparameters = {}

## Photons

The value for photon reflectance and transmission from Zhu et al (2010)

# proportion of photons absorbed by the leaf
modelparameters["P_abs"] = [0.9, r"Dimensionless", "Zhu, Long, and Ort (2010)"]

The value for the thickness of an *Arabidopsis* leaf from Wuyts et al (2021)

modelparameters["T_l"] = [0.017 * 10**-2, r"m", "Wuyts et al. (2010)"]

The are of the leaf being simulated was fixed at 1m$^2$

modelparameters["A_l"] = [1, r"m$^2$", "Fixed"]

The average volume of an *Arabidopsis* guard cell as the midpoint of the values given in table 1 of Jezek and Blatt (2017)

j_b_upper = 0.65
j_b_lower = 0.3
V_gc_ind = (j_b_lower + j_b_upper) / 2  # pL
V_gc_ind = V_gc_ind * 10**-12  # dm3
modelparameters["V_gc_ind"] = [V_gc_ind, r"dm$^3$", "Jezek and Blatt (2017)"]

The photosynthetic efficient of guard cells compared to the mesophyll was taken from Lawson (2003). 

> In leaves of all species the values of photosynthetic efficiency for guard cells were either indistinguishable from or only slightly lower (minimum of 79%) than those of the underlying, spongy mesophyll cells.

I therefore took the average of 100% and 79%, 90%.

modelparameters["FqFm"] = [0.9, r"Dimensionless", "Lawson (2003)"]

The proportion of the number of chloroplasts in guard cells compared to mesophyll cells is taken as the ratio of the midpoint of values for each cell type in Fujiwara et al (2019)

upper_gc = 5.5
lower_gc = 3.5
midpoint_gc = (lower_gc + upper_gc) / 2

upper_me = 100
lower_me = 30
midpoint_me = (upper_me + lower_me) / 2

R_ch = midpoint_gc / midpoint_me

modelparameters["R_ch"] = [R_ch, r"Dimensionless", "Fujiwara, Sanjaya, and Itoh (2019)"]

The proportion of chloroplast volumes between guard cells and mesophyll cells for Arabidopsis Col-0 based on Chl a data from Knoblauch et al. (2023) Table S1

chl_vol_me = 88.24
chl_vol_se_me = 1.58

chl_vol_gc = 17.69
chl_vol_se_gc = 0.21

R_ch_vol = chl_vol_gc / chl_vol_me

modelparameters["R_ch_vol"] = [R_ch_vol, r"Dimensionless", "Knoblauch et al. (2023)"]

The proportion of the leaf that is air from Earles et al (2018) for *Guzmania zahnii* from Table 2. 

modelparameters["L_air"] = [0.37, r"Dimensionless", "Earles et al. (2018)"]

The proportion of the leaf that is epidermis from Wuyts et al 2010

modelparameters["L_epidermis"] = [0.15, r"Dimensionless", "Wuyts et al. (2010)"]

## Osmolarity

Wang 2017 use a value of 0.751 to be the propotion of the Guard Cell that is vacuole

modelparameters["Vac_frac"] = [
    0.751,
    r"Dimensionless",
    "Wang et al. (2017)",
]  # proportion of guard cell that is vacuole

The temperature in the model is taken to be 23C used by Horrer et al 2016. 

modelparameters["T"] = [273.15 + 23, "K", "Horrer et al. (2016)"]

Ideal gas constant, from NIST, (reference Tiesinga2019)

R = 8.205 * 10 ** (-5)  # m3atmK-1mol-1
R = R * 10**3  # dm3atmK-1mol-1
modelparameters["R"] = [R, r"dm$^3$$\cdot$atm$\cdot$K$^{-1}$$\cdot$mol$^{-1}$", "Tiesinga et al. (2019)"]

The density of guard cells per m$^2$ of leaf was taken by multiplying the density of stomata from Papanatsiou et al 2016 by two. This only takes into accound the abaxial surface of the leaf. Strangely, Flutsch et al (2020) had 98 mm$^2$, which is significantly less.

modelparameters["N_gcs"] = [290 * 10**6 * 2.0, r"GCs$\cdot$m$^{-2}$", "Papanatsiou, Amtmann, and Blatt (2016)"]

The following parameters required for osmolarity are based on the updated OnGuard model (Wang et al 2017).

modelparameters["n"] = [2.5, "atm", "Wang et al. (2017)"]
modelparameters["m"] = [0.8, r"atm$\cdot$µm$^{-1}$", "Wang et al. (2017)"]
modelparameters["r"] = [0.05 * 10 ** (-12), r"dm$^3$ µm$^{-1}$", "Wang et al. (2017)"]
modelparameters["s"] = [0.3 * 10 ** (-12), r"dm$^3$", "Wang et al. (2017)"]

apoplastic_concs_wang_2017 = {
    "K": 10,
    "Ca": 1,
    "Cl": 12,
    "Suc": 0.01,
    "MH2": 3.2 * 10 ** (-6),
    "MH": 7.9 * 10 ** (-5),
    "M": 0.00999,
}
apoplastic_conc_wang_2017 = sum(apoplastic_concs_wang_2017.values())

modelparameters["C_apo"] = [
    apoplastic_conc_wang_2017 * 10 ** (-3),
    r"mol$\cdot$dm$^{-3}$",
    "Wang et al. (2017)",
]  # concentration of osmolytes in the apoplast

## Apertures

We use as a source of apertures those from Horrer et al (2016). The list is apertures for the four phases of the model.

modelparameters["A_closed"] = [1.6, r"µm", "Horrer et al. (2016)"]
modelparameters["A_open"] = [2.75, r"µm", "Horrer et al. (2016)"]

modelparameters_df = pd.DataFrame.from_dict(modelparameters, orient="index", columns=["Value", "Units", "Source"])

modelparameters_df

modelparameters_df.to_csv("../inputs/arabidopsis_parameters.csv")
