==> src/1.5_convert_horrer_starch_level.ipynb <==
# Starch amount estimation

Here we use the values determined in Horrer et al. (2016) in order to determine an experimental value for the level of starch degradation we should expect to see in our model. 

import pandas as pd

starchdf = pd.DataFrame([["End of night", 41.92813034], ["1h Light", 12.10368844]],
                        columns=["Time of day", "Starch content per guard cell (pg)"])
starchdf = starchdf.set_index("Time of day")
starchdf

This table is found in the supplementary files of Horrer (2016). We can calculate glucose equivalents from starch by dividing by 162 g/mol, as Horrer et al did:

starchdf["Glucose equivalents (pmol)"] = starchdf["Starch content per guard cell (pg)"]/162
starchdf

In our model we use the glucose equivalents as a measure of starch.

We use a value of $580 \times 10^6$ for the number of guard cells in our model, and therefore we can calculate the expected amount of starch in the model at the two times, and the expected amount of degradation.

N_gc = 290 * 10 ** 6 * 2  # number of stomata/m^2 *2 gives number of guard cells
starchdf["Starch amount (mmole/m2)"] = starchdf["Glucose equivalents (pmol)"]*N_gc*10**-9  # 10**-9 to convert to mmol
starchdf

print("Expected starch degradation amount is therefore %s fmoles/GC" %
      round((starchdf.at["End of night", "Glucose equivalents (pmol)"]-starchdf.at["1h Light", "Glucose equivalents (pmol)"])*1000, 3))

print("Expected starch degradation amount is therefore %s mmoles/m2" %
      round((starchdf.at["End of night", "Starch amount (mmole/m2)"]-starchdf.at["1h Light", "Starch amount (mmole/m2)"]), 3))
