==> src/3.1_generate_constraints_df.ipynb <==
# Constraint scan

## Generate constraintscandf

import numpy as np
import pandas as pd
import pyDOE
from mmon_gcm.constraintscan import generate_constraints_df

## Define parameter bounds

parameter_bounds = {}

## Photons

We can go 10% above and below the value for photon reflectance and transmission from Zhu et al (2010)

parameter_bounds["P_abs"] = [0.9*0.9, 0.9*1.1, r"Dimensionless",
                             "90% of Zhu, Long, and Ort (2010)", "110% of Zhu, Long, and Ort (2010)"]  # proportion of photons absorbed by the leaf

The lower bound is the default we've used from Wuyts et al (2021), for a higher value we'll use one from Ramonell 2001.

parameter_bounds["T_l"] = [0.017 * 10 ** -2, 0.024 * 10 ** -2, r"m", "Wuyts et al. (2010)", "Ramonell et al. (2001)"]

The are of the leaf being simulated is fixed at 1m$^2$

parameter_bounds["A_l"] = [1, 1, r"m$^2$", "Fixed", "Fixed"]

For the lower bound we can use the average volume of an *Arabidopsis* guard cell as the midpoint of the values given in table 1 of Jezek and Blatt (2017). For the upper bound we can use the volume given in Hills et al (2012). 

j_b_upper = 0.65
j_b_lower = 0.3
V_gc_ind = (j_b_lower+j_b_upper)/2  # pL
V_gc_ind = V_gc_ind * 10**-12  # dm3
parameter_bounds["V_gc_ind"] = [V_gc_ind, 4.1E-12, r"dm$^3$", "Jezek and Blatt (2017)", "Hills et al. (2012)"]

The photosynthetic efficient of guard cells compared to the mesophyll was taken from Lawson (2003). 

> In leaves of all species the values of photosynthetic efficiency for guard cells were either indistinguishable from or only slightly lower (minimum of 79%) than those of the underlying, spongy mesophyll cells.

We'll use these as the bounds.

parameter_bounds["FqFm"] = [0.79, 0.9, r"Dimensionless", "Lawson (2003)", "Lawson (2003)"]

Fujiwara et al (2019) provide a range of values for number of chloroplasts in mesophyll and guard cells, we can take the upper gc and lower me and vice versa to get the upper and lower bound for the ratios, respectively.

upper_gc = 5.5
lower_gc = 3.5

upper_me = 100
lower_me = 30

rch_lower_bound = lower_gc/upper_me
rch_upper_bound = upper_gc/lower_me

parameter_bounds["R_ch"] = [rch_lower_bound, rch_upper_bound, r"Dimensionless",
                            "Fujiwara, Sanjaya, and Itoh (2019)", "Fujiwara, Sanjaya, and Itoh (2019)"]

Added in Nov 2023: The proportion of chloroplast volumes between guard cells and mesophyll cells for Arabidopsis Col-0 based on Chl a data from Knoblauch et al. (2023) Table S1. We can the standard error values to calculate the range with min/max being 2 SE away.

chl_vol_me = 88.24
chl_vol_se_me = 1.58
min_chl_vol_me = chl_vol_me - 2 * chl_vol_se_me
max_chl_vol_me = chl_vol_me + 2 * chl_vol_se_me

chl_vol_gc = 17.69
chl_vol_se_gc = 0.21
min_chl_vol_gc = chl_vol_gc - 2 * chl_vol_se_gc
max_chl_vol_gc = chl_vol_gc + 2 * chl_vol_se_gc

min_R_ch_vol = min_chl_vol_gc / max_chl_vol_me
max_R_ch_vol = max_chl_vol_gc / min_chl_vol_me

#modelparameters["R_ch_vol"] = [R_ch_vol, r"Dimensionless", "Knoblauch et al. (2023)"]

parameter_bounds["R_ch_vol"] = [min_R_ch_vol, max_R_ch_vol, r"Dimensionless",
                            "Knoblauch et al. (2023)", "Knoblauch et al. (2023)"]

For the lower bound Ramonell et al. (2001) provide a percentage of the leaf that is space at atmospheric pressure in Table 1.For the upper bound we take the proportion of the leaf that is air from Earles et al 2018 for *Guzmania zahnii* from Table 2. 

parameter_bounds["L_air"] = [0.185, 0.37, r"Dimensionless", "Ramonell et al. (2001)", "Earles et al. (2018)"]

Willmer and Fricker (1996) provide a lower bound of 0.1 and Ramonell et al. (2001) an upper bound of 0.24 for the proportion of the leaf that is epidermis.

parameter_bounds["L_epidermis"] = [0.1, 0.24, r"Dimensionless", "Willmer and Fricker (1996)", "Ramonell et al. (2001)"]

## Osmolarity

Wang et al. (2017) use a value of 0.751 to be the proportion of the Guard Cell that is vacuole, Andrés et al. (2014) quote a value of 90% from an old MacRobbie paper.  

parameter_bounds["Vac_frac"] = [0.751, 0.9, r"Dimensionless", "Wang et al. (2017)", "Andrés et al. (2014) "]

A reasonable range for temperature seems to be 10C to 25C, this covers a wide range:

parameter_bounds["T"] = [273.15+10, 273.15+25, "K", "10C", "25C"]

Ideal gas constant, from NIST, (reference Tiesinga2019). This is constant, so won't change in the model

R = 8.205*10**(-5)  # m3atmK-1mol-1
R = R*10**3  # dm3atmK-1mol-1
parameter_bounds["R"] = [
    R, R, r"dm$^3$$\cdot$atm$\cdot$K$^{-1}$$\cdot$mol$^{-1}$", "Tiesinga et al. (2019)", "Tiesinga et al. (2019)"]

The density of guard cells per m$^2$ of leaf was taken by multiplying the density of stomata from Papanatsiou et al (2016) by two. This only takes into account the abaxial surface of the leaf, so for the upper bound, we'll double this again. For the lower bound, we'll use a value for *Commelina* from Willmer and Fricker (1996). 

parameter_bounds["N_gcs"] = [172 * 10 ** 6, 290 * 10 ** 6 * 2.0 * 2,
                             r"GCs$\cdot$m$^{-2}$", "Willmer and Fricker (1996)", "Papanatsiou, Amtmann, and Blatt (2016)"]

The following parameters required for osmolarity are based taking values from an older OnGuard paper (Wang et al. (2012)) and the newer, updated model (Wang et al 2017).

parameter_bounds["n"] = [1.5, 2.5, "atm", "Wang et al. (2012)", "Wang et al. (2017)"]
parameter_bounds["m"] = [0.8, 1, r"atm$\cdot$µm$^{-1}$", "Wang et al. (2017)", "Wang et al. (2012)"]
parameter_bounds["r"] = [0.05 * 10 ** (-12), 0.08 * 10 ** (-12),
                         r"dm$^3$ µm$^{-1}$", "Wang et al. (2017)", "Wang et al. (2012)"]
parameter_bounds["s"] = [0.1 * 10 ** (-12), 0.3 * 10 ** (-12), r"dm$^3$", "Wang et al. (2012)", "Wang et al. (2017)"]

For the apoplastic concentration, we use the value from the OnGuard model as the lower bound, and use a combination of values which have been collated by Roelfsema and Hedrich (2002) from Lohaus et al. (2001) for the upper bound:

apoplastic_concs_wang_2017 = {
    "K": 10,
    "Ca": 1,
    "Cl": 12,
    "Suc": 0.01,
    "MH2": 3.2*10**(-6),
    "MH": 7.9*10**(-5),
    "M": 0.00999,
}
apoplastic_conc_wang_2017 = sum(apoplastic_concs_wang_2017.values())

apoplastic_concs_roelfsema_2002 = {
    "K": 13,
    "Ca": 0.7,
    "Cl": 11,
    "Suc": 1.6,
    "Mal": 0.7,
    "Amino Acids": 9.6,
    "Hexoses": 0.7,
}
apoplastic_conc_roelfsema_2002 = sum(apoplastic_concs_roelfsema_2002.values())

parameter_bounds["C_apo"] = [apoplastic_conc_wang_2017*10 **
                             (-3), apoplastic_conc_roelfsema_2002*10**(-3), r"mol$\cdot$dm$^{-3}$", "Wang et al. (2017)", "Roelfsema and Hedrich (2002)"]

## Apertures

For the closed aperture, we'll use 1 µm value from Jezek and Blatt (2017) as a lower bound and 4 µm from Wang et al. (2017) as an upper bound.

parameter_bounds["A_closed"] = [1, 4, r"µm", "Jezek and Blatt (2017)", "Wang et al. (2017)"]

For the open aperture, we'll use the 2.75 µm from Horrer et al. (2016) as the lower bound and 12 µm from Wang et al. (2017) as the upper bound.

parameter_bounds["A_open"] = [2.75, 12, r"µm", "Horrer et al. (2016)", "Wang et al. (2017)"]

For the proportion of photons that are absorbed by the leaf we can just go for 10% above and below the value that is reported by 

## ATPase

We'll go between constraining the ATPase to 0 and the level of ATPase that Flütsch et al. (2020) measured (17 fmoles$\cdot$GC$^{-1}\cdot$h$^{-1}$).

parameter_bounds["ATPase"] = [
    0, 17, r"fmoles$\cdot$GC$^{-1}\cdot$h$^{-1}$", "Supplementary", "Flütsch et al. (2020)"]

# Export parameters to csv

params_df = pd.DataFrame.from_dict(parameter_bounds, orient="index", columns=[
                                   "Lower", "Upper", "Units", "Source Lower", "Source Upper"])
params_df.to_csv("../outputs/constraint_scan/parameter_bounds.csv")
params_df.to_csv("../plant_cell_paper/supplemental_table_2.csv")

params_df

### Generate constraints array

np.random.seed(12345)
lhs_df = pd.DataFrame(pyDOE.lhs(len(params_df), samples=1000), columns=params_df.index)

lhs_df

constraints_df = generate_constraints_df(lhs_df, params_df)
constraints_df.head()

constraints_df.to_csv("../outputs/constraint_scan/constraints_df.csv")
